Revert "feat : clean functions in heredoc #38"
This reverts commit 1b976765699b8bd7ed65b98111a6e51de6c20626.
