+++ jaemjeon_jisookim_minishell/main/Makefile
@@ -52,7 +52,7 @@ SRC_DEBUG		:= $(addprefix $(DIR_DEBUG)/, $(SRC_DEBUG))
 
 SRC_EXECUTE		:=	exec_main.c exec_multi_cmd.c exec_tools.c\
 					exec_tools_file.c exec_redirection.c \
-					exec_check_built_in.c
+					exec_redi_parse.c exec_check_built_in.c
 SRC_EXECUTE		:=	$(addprefix $(DIR_EXECUTE)/, $(SRC_EXECUTE))
 
 SRC_EXE_INIT	:=	exec_init_double_env.c exec_init_struct.c\
@@ -63,7 +63,7 @@ SRC_BUILT_IN	:=	ft_cd.c ft_echo.c ft_env.c ft_exit.c ft_export.c ft_pwd.c \
 					ft_unset.c
 SRC_BUILT_IN	:=	$(addprefix $(DIR_BUILT_IN)/, $(SRC_BUILT_IN))
 
-SRC_HEREDOC		:=	heredoc.c heredoc_init.c heredoc_tools.c
+SRC_HEREDOC		:=	heredoc.c heredoc_init.c heredoc_check.c
 SRC_HEREDOC		:=	$(addprefix $(DIR_HEREDOC)/, $(SRC_HEREDOC))
 
 SRCS			=	$(SRC_ENV) $(SRC_TOKENIZE) $(SRC_UTILS) $(SRC_MAIN) \
