+++ jaemjeon_jisookim_minishell/main/include/heredoc.h
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 14:49:50 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/18 01:56:57 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/17 19:00:24 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,21 +28,18 @@ typedef struct s_hdoc
 	char		**file_name;
 }	t_hdoc;
 
-//heredoc_tools
-void	doing_heredoc_sigint_process(int signo);
-void	heredoc_process_sigint_process(int signo);
-char	*make_tmp_filename(void *p1_8byte, void *p2_8byte);
-void	rename_string_value(t_exec *exec);
+
+// heredoc_check
 
 //heredoc
-int		make_heredoc_file(t_exec *exec, t_cmd *cur_cmd, \
-												t_token *cur_redirect_token);
 int		heredoc(t_exec *exec, pid_t ret_pid);
+void	make_and_open_hdoc(t_exec *exec, t_hdoc *hdoc, char *limiter, int i);
 void	do_heredoc(t_exec *exec, char *limiter, int fd);
+void	close_and_reopen_hdoc(t_exec *exec, t_hdoc *hdoc, int i);
 
 // heredoc_init
 void	count_hdoc(t_exec *exec, t_hdoc *hdoc);
-void	make_heredoc_lilmiter_array(t_exec *exec, t_hdoc *hdoc);
 int		init_info_hdoc_struct(t_exec *exec, t_hdoc *hdoc);
+void	make_heredoc_lilmiter_array(t_exec *exec, t_hdoc *hdoc);
 
 #endif
