+++ jaemjeon_jisookim_minishell/main/src/execute/exec_main.c
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 15:15:40 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/18 01:35:52 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/18 01:15:07 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
