+++ jaemjeon_jisookim_minishell/main/src/execute/exec_tools.c
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 15:15:35 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/18 01:43:00 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/18 00:55:13 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,7 +87,9 @@ void	exec_executing(t_exec *exec, int process_number)
 		exit(0);
 	if (stat == -1)
 	{
-		
+		ft_putstr_fd(strerror(errno), 2);
+		ft_putstr_fd("\n", 2);
+		exit(1);
 	}
 }
